+++ Assets/Scripts/Data/Models/IGameModel.cs
@@ -17,7 +17,7 @@ namespace BattleCruisers.Data.Models
         int CreditsChange { get; set; }
         string PlayerName { get; set; }
         int GameMap { get; set; }
-        int BattleWinScore { get; set; }
+        float BattleWinScore { get; set; }
         string QueueName { get; set; }
         List<Arena> Arenas { get; set; }
         Dictionary<string, int> GameConfigs { get; set; }
